fix: correct ray-AABB bounding-box to NAADF's 0.1-voxel-inset float3 values; document Phase A-2 TAA context
- `GpuWorldMeta.bounding_box_min/max` changed from `IVec3` (integer-inclusive) to `Vec3` (float), matching NAADF's `float3 boundingBoxMin/Max` in `rayTracing.fxh`.
- `prepare.rs`: upload `min = (0.1,0.1,0.1)` and `max = sizeInVoxels - 0.1`, replicating `WorldData.cs:477-478` exactly; fixes out-of-volume concentric-line artifact caused by `floor()` flipping on integer voxel boundaries.
- `ray_tracing.wgsl`: `shoot_ray` lower-bound break now tests `floor(bounding_box_min)` so valid edge cells (index 0) are not evicted by the 0.1 inset.
- `world_data.wgsl` / `naadf_first_hit.wgsl`: `bbox_min/max` bindings retype to `vec3<f32>`, no cast needed.
- `05-review.md`: full diagnose-and-fix record for the second Phase-A review-gate bug.
- `01-context.md`: Phase A-2 (TAA) working context — 16-sample history, 0.25-spp GI target constraint, scope, done-bar.
- `README.md`: mark Phase A review gate passed; mark Phase A-2 design in progress.

+++ src/assets/shaders/world_data.wgsl
@@ -30,10 +30,13 @@
 struct GpuWorldMeta {
     // World size in chunks.
     size_in_chunks: vec3<u32>,
-    // Geometry AABB minimum, in voxels (inclusive).
-    bounding_box_min: vec3<i32>,
-    // Geometry AABB maximum, in voxels (inclusive).
-    bounding_box_max: vec3<i32>,
+    // Geometry AABB minimum, in voxels — NAADF's `boundingBoxMin` (the
+    // 0.1-voxel-inset world minimum, `WorldData.cs:477`). `float3`, not
+    // integer, faithful to `rayTracing.fxh`'s `float3 boundingBoxMin`.
+    bounding_box_min: vec3<f32>,
+    // Geometry AABB maximum, in voxels — NAADF's `boundingBoxMax`
+    // (`sizeInVoxels - 0.1`, `WorldData.cs:478`).
+    bounding_box_max: vec3<f32>,
 }
 
 // --- @group(0) — world data (read-only in the render passes) ----------------
